Update stores_info.csv
I looked into the stores that had 0 crimes. Most of them were located in different cities, and I confirmed that those cities have their own police departments. There were 2 others that do seem to match up to our crime list, but the address was far enough off that the code wasn't picking it up. For this set, 100% have crimes associated with them. I'm going to also take a look at some of the ones that have really low numbers to see if they might have different forms of the addresses that we're not picking up.